New ElcanoSerial Optimizations
